Refactor + attemps to run bot with server both
